+++ server/application/controllers/api/pelanggan.php
@@ -2,6 +2,7 @@
 
 defined('BASEPATH') or exit('No direct script access allowed');
 require APPPATH . '/libraries/REST_Controller.php';
+require_once FCPATH . 'vendor/autoload.php';
 
 use Restserver\Libraries\REST_Controller;
 
@@ -19,6 +20,7 @@ class Pelanggan extends REST_Controller
         $this->load->database();
         $this->load->model('M_Pelanggan');
         $this->load->library('form_validation');
+        $this->load->library('jwt');
     }
 
     function validate(){
@@ -32,8 +34,6 @@ class Pelanggan extends REST_Controller
         $this->form_validation->set_rules('no_hp', 'Nomor Hp Pelanggan', 'required|trim');
     }
 
-
-
     public function options_get() {
         header("Access-Control-Allow-Origin: *");
         header("Access-Control-Allow-Methods: GET, POST, OPTIONS, PUT, DELETE");
@@ -41,18 +41,27 @@ class Pelanggan extends REST_Controller
         exit();
     }
 
-    function index_get() {
-        // $authorizationHeader = $this->input->get_request_header('Authorization', true);
+    function is_login() {
+        $authorizationHeader = $this->input->get_request_header('Authorization', true);
 
-        // if(empty($authorizationHeader) || $this->jwt->decode($authorizationHeader) === false) {
-        //     return $this->response(
-        //         array(
-        //             'kode' => '401',
-        //             'pesan' => 'signature tidak sesuai',
-        //             'data' => []
-        //         ), '401'
-        //     );
-        // }
+        if (empty($authorizationHeader) || $this->jwt->decode($authorizationHeader) === false) {
+            $this->response(
+                array(
+                    'kode' => '401',
+                    'pesan' => 'signature tidak sesuai',
+                    'data' => []
+                ), '401'
+            );
+            return false;
+        }
+
+        return true;
+    }
+
+    function index_get() {
+        if (!$this->is_login()) {
+            return;
+        }
 
         $id = $this->get('id');
         if($id == '') {
@@ -64,6 +73,10 @@ class Pelanggan extends REST_Controller
     }
     
     function index_post(){
+        if (!$this->is_login()) {
+            return;
+        }
+
         $this->validate();
         if ($this->form_validation->run() === FALSE) {
             $error = $this->form_validation->error_array();
@@ -92,9 +105,11 @@ class Pelanggan extends REST_Controller
             return $this->response($response);
     }
 
-
-    // ERROR UPDATE DATA FORM VALIDATION
     public function index_put(){
+        if (!$this->is_login()) {
+            return;
+        }
+
         $id = $this->put('id');
         $check = $this->M_Pelanggan->check_data($id);
 
@@ -137,6 +152,10 @@ class Pelanggan extends REST_Controller
     }
     
     function index_delete() {
+        if (!$this->is_login()) {
+            return;
+        }
+
         $id = $this->delete('id');
         $check = $this->M_Pelanggan->check_data($id);
         if($check == false) {
